+++ ui/src/routes/methodology.svelte
@@ -34,32 +34,89 @@
 		<h1>Methodology</h1>
 
 		<h2>Data sources</h2>
-
-		<p>As much as possible we have used data from official sources such as <a href={onsUrl}>ONS</a>, <a href={eurostatUrl}>Eurostat</a>, <a href={hesaUrl}>HESA</a> or <a href={ukriUrl}>UKRI</a>. One reason to do this is to enable the reproducibility of our analysis, and to remove the reliance of the tool on proprietary sources.<p>
-		<p>Having said this, in a small number of instances we have used proprietary data sources such as <a href={patstatUrl}>PATSTAT</a> for the analysis of patenting, and <a href={crunchbaseUrl}>Crunchbase</a> for the analysis of venture capital investment.<p>
+		<p>
+			As much as possible we have used data from official sources such as
+			<a href={onsUrl} rel='noopener'>ONS</a>,
+			<a href={eurostatUrl} rel='noopener'>Eurostat</a>,
+			<a href={hesaUrl} rel='noopener'>HESA</a> or
+			<a href={ukriUrl} rel='noopener'>UKRI</a>. One reason to do this is
+			to enable the reproducibility of our analysis, and to remove the
+			reliance of the tool on proprietary sources.
+		</p>
+		<p>
+			Having said this, in a small number of instances we have used
+			proprietary data sources such as
+			<a href={patstatUrl} rel='noopener'>PATSTAT</a> for the analysis of
+			patenting, and <a href={crunchbaseUrl} rel='noopener'>Crunchbase</a>
+			for the analysis of venture capital investment.
+		</p>
 
 		<h2>Geographies</h2>
-
-		<p>We use <a href={nutsUrl}>NUTS2</a> regions as our geographical unit of analysis. This has allowed us to collect data about regional R&D activity which is only available at that level. We note that were possible we have also calculated indicators at a higher level of granularity (NUTS3) as well as using policy-relevant <a href={lepUrl}>LEP</a>s boundaries. These will be released when the tool is published later in 2020.<p>
-		<p>In many cases we have reverse geocoded observations available at high level of geographical resolution (for example, the geographical coordinates of a higher education institution) using NUTS2 boundary files available from Eurostat. When doing this, we have assigned observations to regions in the NUTS2 version that was in use at the time when the data were collected / when the events captured in the data took place.<p>
+		<p>
+			We use <a href={nutsUrl} rel='noopener'>NUTS2</a> regions as our
+			geographical unit of analysis. This has allowed us to collect data
+			about regional R&D activity which is only available at that level.
+			We note that were possible we have also calculated indicators at a
+			higher level of granularity (NUTS3) as well as using policy-relevant
+			<a href={lepUrl} rel='noopener'>LEP</a>s boundaries. These will be
+			released when the tool is published later in 2020.
+		</p>
+		<p>
+			In many cases we have reverse geocoded observations available at
+			high level of geographical resolution (for example, the geographical
+			coordinates of a higher education institution) using NUTS2 boundary
+			files available from Eurostat. When doing this, we have assigned
+			observations to regions in the NUTS2 version that was in use at the
+			time when the data were collected / when the events captured in the
+			data took place.
+		</p>
 
 		<h2>Data processing</h2>
-
-		<p>In general, we have avoided complex data processing beyond what was required to aggregate data at our preferred level of geographical resolution. There are however a couple of exceptions to this:<p>
+		<p>
+			In general, we have avoided complex data processing beyond what was
+			required to aggregate data at our preferred level of geographical
+			resolution. There are however a couple of exceptions to this:
+		</p>
 
 		<ul>
 			<li>
-				<p>We have calculated indices of economic complexity for UK regions used the algorithm developed by <ExternalLink href={HausmanAndHidalgoUrl} text='Hausman and Hidalgo (2009)'/>.<p>
+				<p>
+					We have calculated indices of economic complexity for UK
+					regions used the algorithm developed by
+					<ExternalLink
+						href={HausmanAndHidalgoUrl}
+						text='Hausman and Hidalgo (2009)'
+					/>.
+				</p>
 			</li>
 			<li>
-				<p>We have measured levels of employment in entertainment and cultural sectors using an industrial segmentation based on the methodology developed by <ExternalLink href={DelgadoEtAlUrl} text='Delgado et al (2015)'/>.<p>
+				<p>
+					We have measured levels of employment in entertainment and
+					cultural sectors using an industrial segmentation based on
+					the methodology developed by
+					<ExternalLink
+						href={DelgadoEtAlUrl}
+						text='Delgado et al (2015)'
+					/>.
+				</p>
 			</li>
 			<li>
-				<p>We have identified UKRI-funded research projects in STEM disciplines using a machine learning analysis of project descriptions presented in <ExternalLink href={MateosGarciaUrl} text='Mateos-Garcia (2017)'/>.<p>
+				<p>
+					We have identified UKRI-funded research projects in STEM
+					disciplines using a machine learning analysis of project
+					descriptions presented in
+					<ExternalLink
+						href={MateosGarciaUrl}
+						text='Mateos-Garcia (2017)'
+					/>.
+				</p>
 			</li>
 		</ul>
 
-		<p>We indicate those indicators based on experimental methodologies or data sources where relevant.<p>
+		<p>
+			We indicate those indicators based on experimental methodologies or
+			data sources where relevant.
+		</p>
 
 		<div class='cta'>
 			<LinkButton
@@ -86,7 +143,6 @@
 	main {
 		background-color: var(--color-background);
 		display: flex;
-		font-size: 1.05rem;
 		font-weight: 200;
 		height: 100%;
 		justify-content: space-around;
@@ -102,10 +158,10 @@
 	}
 
 	h1 {
-		font-family: 'Open Sans Semibold', sans-serif;
+		font-weight: bold;
 	}
 	h2 {
-		font-family: 'Open Sans Regular', sans-serif;
+		font-weight: normal;
 		margin-bottom: 1.5rem;
 		margin-top: 1.5rem;
 	}
@@ -115,16 +171,16 @@
 	}
 
 	a {
-		font-weight: bold;
 		text-decoration: none;
+		font-weight: bold;
 	}
 
 	p a {
+		padding: 0.1rem;
 		border-bottom: 1px solid var(--color-link);
 		color: var(--color-link);
-		font-weight: bold;
-		padding: 0.1rem;
 		text-decoration: none;
+		font-weight: bold;
 	}
 
 	ul {