+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Make the app skeleton responsive (#448)
 - Add the accessibility menu (#446)
 - Set gov.uk blue as main color (#455, #457)
+- Format text routes (#447)
 
 # 0.0.17 (staging release)